Update Upsert to return error
This commit updates the Upsert method to return just an error
instead of the previous (string, error). This brings it up to
date with changes upstream (#3881) where we removed the string
return value because it was redundant. If error is non-nil,
then the id is assigned to the *pb.Silence and does not need
to be returned.