Update Upsert to use latest fixes
This commit updates the Upsert method to use the latest fixes
from upstream. It must set the UpdatedAt timestamp before
calling setSilence as this code was removed from setSilence,
and it must also change the silence into a *pb.MeshSilence
before calling setSilence as it now expects a *pb.MeshSilence
instead of *pb.Silence.